+++ cli/src/inspect.rs
@@ -101,7 +101,7 @@ pub fn relation(
 }
 
 fn column_type(column: &elephantry::inspect::Column) -> String {
-    let ty = if let Ok(ty) = elephantry::pq::types::Type::try_from(column.oid) {
+    if let Ok(ty) = elephantry::pq::types::Type::try_from(column.oid) {
         match ty.kind {
             elephantry::pq::types::Kind::Array(_) => {
                 format!("{}[]", ty.name.trim_start_matches('_'))
@@ -110,9 +110,7 @@ fn column_type(column: &elephantry::inspect::Column) -> String {
         }
     } else {
         column.ty.clone()
-    };
-
-    ty
+    }
 }
 
 pub fn enums(connection: &elephantry::Connection, schema: &str) -> crate::Result {