+++ core/src/inspect.rs
@@ -112,9 +112,7 @@ select c.oid as oid
     "#,
             &[&schema, &relation],
         )
-        .map_err(|_| {
-            return crate::Error::Inspect(format!("Unknow relation {schema}.{relation}"));
-        })?;
+        .map_err(|_| crate::Error::Inspect(format!("Unknow relation {schema}.{relation}")))?;
 
     connection
         .query(
@@ -277,7 +275,5 @@ where s.nspname = $*
     "#,
             &[&name],
         )
-        .map_err(|_| {
-            return crate::Error::Inspect(format!("Unknow schema {name}"));
-        })
+        .map_err(|_| crate::Error::Inspect(format!("Unknow schema {name}")))
 }