+++ core/src/sql/jsonb.rs
@@ -1,7 +1,7 @@
 /**
  * Rust type for jsonb.
  */
-#[derive(Clone, Debug, PartialEq)]
+#[derive(Clone, Debug, Eq, PartialEq)]
 #[cfg_attr(docsrs, doc(cfg(feature = "json")))]
 pub struct Jsonb(serde_json::Value);